The merge extracts common additions before an actual conflict
This is useful as it makes the conflicts smaller:

  +common_line
  <<<<<
  -old_line
  +line_one
  >>>>>
  -old_line
  +line_two
  =====

instead of

  <<<<<
  +common_line
  -old_line
  +line_one
  >>>>>
  +common_line
  -old_line
  +line_two
  =====

The '-old_line' is part of the change itself. A good interface should print this outside of the '<<<<<' and '>>>>>'

However, there is another case which is not covered yet:

  <<<<<
  +line_one
  common_line
  >>>>>
  +line_two
  common_line
  =====

In this case, the common_line should be extracted as a separate change.